PopulateDatabase/requirements.txt
requests
mysql-connector-python
faker